add version to types
