+++ client/query.go
@@ -5,7 +5,8 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/EspressoSystems/espresso-sequencer-go/types"
+	types "github.com/EspressoSystems/espresso-sequencer-go/types"
+	common "github.com/EspressoSystems/espresso-sequencer-go/types/common"
 )
 
 // Interface to the Espresso Sequencer query service.
@@ -20,23 +21,23 @@ type QueryService interface {
 	// along with a proof that these are all such transactions.
 	FetchTransactionsInBlock(ctx context.Context, blockHeight uint64, namespace uint64) (TransactionsInBlock, error)
 	// Get the transaction by its hash.
-	FetchTransactionByHash(ctx context.Context, hash *types.TaggedBase64) (types.TransactionQueryData, error)
+	FetchTransactionByHash(ctx context.Context, hash *common.TaggedBase64) (common.TransactionQueryData, error)
 }
 
 // Response to `FetchTransactionsInBlock`
 type TransactionsInBlock struct {
 	// The transactions.
-	Transactions []types.Bytes `json:"transactions"`
+	Transactions []common.Bytes `json:"transactions"`
 	// A proof that these are all the transactions in the block with the requested namespace.
-	Proof     types.NamespaceProof `json:"proof"`
-	VidCommon types.VidCommon      `json:"vidCommon"`
+	Proof     common.NamespaceProof `json:"proof"`
+	VidCommon common.VidCommon      `json:"vidCommon"`
 }
 
 func (t *TransactionsInBlock) UnmarshalJSON(b []byte) error {
 	// Parse using pointers so we can distinguish between missing and default fields.
 	type Dec struct {
-		Transactions *[]types.Bytes        `json:"transactions"`
-		Proof        *types.NamespaceProof `json:"proof"`
+		Transactions *[]common.Bytes        `json:"transactions"`
+		Proof        *common.NamespaceProof `json:"proof"`
 	}
 
 	var dec Dec
