+++ client/submit.go
@@ -3,11 +3,11 @@ package client
 import (
 	"context"
 
-	"github.com/EspressoSystems/espresso-sequencer-go/types"
+	common "github.com/EspressoSystems/espresso-sequencer-go/types/common"
 )
 
 // Interface to the Espresso Sequencer submit API
 type SubmitAPI interface {
 	// Submit a transaction to the espresso sequencer.
-	SubmitTransaction(ctx context.Context, tx types.Transaction) (*types.TaggedBase64, error)
+	SubmitTransaction(ctx context.Context, tx common.Transaction) (*common.TaggedBase64, error)
 }
