+++ types/common/commit.go
@@ -1,4 +1,4 @@
-package types
+package common
 
 import (
 	"bytes"
