+++ types/common/consts.go
@@ -1,4 +1,4 @@
-package types
+package common
 
 func GetMagicBytes() [32]byte {
 	return NewRawCommitmentBuilder("espresso-builder-zNC8sXSk5Yl6Uiu").Finalize()
